`KafkaConsumerConfig` add `SwiftKafka` options (#54)
* `KafkaConsumerConfig` add `SwiftKafka` options

Motivation:

We want to use the `KafkaConsumerConfig` to determine if the
`KafkaConsumer` is part of a consumer-group (subscription) or assinged
to a single partition of a topic (assingment). This should also be
implemented in a way that errors are caught at compile time.
Furthermore, we want users to be able to determine a
`BackPressureStrategy` for their respective `KafkaClient`.

Modifications:

* add two new configuration options to `KafkaConsumerConfig`:
  `ConsumptionStrategy` and `BackPressureStrategy`
* remove `KafkaConsumerConfig`'s `groupID` option, this is now taken
  care of by the `consumptionStrategy` option, which is better as this
  ensures at compile time that now groupID mismatch errors occur
* rename `KafkaProducerConfig` back to `ConfigEnums` as it is not really
  meant for shared configurations, rather as a namespace for _all_
  configuration options
* update README

* * move setting group.id to `KafkaConsumerConfig`

* * fix soundness

* * rename `ConfigEnums` -> `KafkaSharedConfiguration`

* Review Franz

Modifications:

* fix wrong Docc comments
* remove implicitly unwrapped optional
* rename `highLowWatermark(lowWatermark:,highWatermark:)` to
  `watermark(low:,high:)`
* `ConsumptionStrategy`: remove *Based suffixes
* replace type of `KafkaConsumer.offset`: `Int64` -> `Int`

* * fix docc nits

+++ Sources/SwiftKafka/KafkaAcknowledgedMessage.swift
@@ -30,7 +30,7 @@ public struct KafkaAcknowledgedMessage: Hashable {
     /// The body of the message.
     public var value: ByteBuffer
     /// The offset of the message in its partition.
-    public var offset: Int64
+    public var offset: Int
 
     /// Initialize ``KafkaAcknowledgedMessage`` from `rd_kafka_message_t` pointer.
     /// - Throws: A ``KafkaAcknowledgedMessageError`` for failed acknowledgements or malformed messages.
@@ -63,6 +63,6 @@ public struct KafkaAcknowledgedMessage: Hashable {
             self.key = nil
         }
 
-        self.offset = Int64(rdKafkaMessage.offset)
+        self.offset = Int(rdKafkaMessage.offset)
     }
 }
